# openapi.yaml
openapi: 3.1.0
info:
  title: RescueBite API
  description: Production-grade RescueBite backend with restaurant operations, inventory
    expiry, reservations, and manual seller-confirmed payments.
  version: 1.1.0
paths:
...
  /users/me:
    get:
      tags:
      - users
      summary: Me
      operationId: me_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Standard 422 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - users
      summary: Update Me
      operationId: update_me_users_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /users/me/addresses:
    get:
      tags:
      - users
      summary: List Addresses
      operationId: list_addresses_users_me_addresses_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AddressRead'
                type: array
                title: Response List Addresses Users Me Addresses Get
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Standard 422 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - users
      summary: Create Address
      operationId: create_address_users_me_addresses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /users/me/addresses/{addressId}:
    patch:
      tags:
      - users
      summary: Update Address
      operationId: update_address_users_me_addresses__addressId__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: addressId
        in: path
        required: true
        schema:
          type: integer
          title: Addressid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - users
      summary: Delete Address
      operationId: delete_address_users_me_addresses__addressId__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: addressId
        in: path
        required: true
        schema:
          type: integer
          title: Addressid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/allergies:
    put:
      tags:
      - users
      summary: Replace Allergies
      operationId: replace_allergies_users_me_allergies_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllergyUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllergyUpdate'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
    delete:
      tags:
      - users
      summary: Clear Allergies
      operationId: clear_allergies_users_me_allergies_delete
      responses:
        '204':
          description: Successful Response
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Standard 422 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /users/me/allergies/check:
    post:
      tags:
      - users
      summary: Check Ingredient Payload
      operationId: check_ingredient_payload_users_me_allergies_check_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllergyCheckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllergyCheckResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
...
components:
  schemas:
    AddressCreate:
      properties:
        label:
          type: string
          maxLength: 80
          title: Label
        city:
          type: string
          title: City
        street:
          type: string
          title: Street
        lat:
          type: number
          title: Lat
        lng:
          type: number
          title: Lng
      type: object
      required:
      - label
      - city
      - street
      - lat
      - lng
      title: AddressCreate
    AddressRead:
      properties:
        label:
          type: string
          maxLength: 80
          title: Label
        city:
          type: string
          title: City
        street:
          type: string
          title: Street
        lat:
          type: number
          title: Lat
        lng:
          type: number
          title: Lng
        id:
          type: integer
          title: Id
      type: object
      required:
      - label
      - city
      - street
      - lat
      - lng
      - id
      title: AddressRead
    Allergen:
      type: string
      enum:
      - peanuts
      - tree_nuts
      - dairy
      - eggs
      - gluten
      - soy
      - fish
      - shellfish
      - sesame
      title: Allergen
    AllergyCheckRequest:
      properties:
        ingredients:
          items:
            $ref: '#/components/schemas/IngredientIn'
          type: array
          title: Ingredients
      type: object
      required:
      - ingredients
      title: AllergyCheckRequest
    AllergyCheckResponse:
      properties:
        safe:
          type: boolean
          title: Safe
        conflicts:
          items:
            $ref: '#/components/schemas/Allergen'
          type: array
          title: Conflicts
      type: object
      required:
      - safe
      - conflicts
      title: AllergyCheckResponse
    AllergyUpdate:
      properties:
        allergens:
          items:
            $ref: '#/components/schemas/Allergen'
          type: array
          title: Allergens
      type: object
      required:
      - allergens
      title: AllergyUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IngredientIn:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        allergen:
          anyOf:
          - $ref: '#/components/schemas/Allergen'
          - type: 'null'
      type: object
      required:
      - name
      title: IngredientIn
...
    UserRead:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          type: string
          title: Status
      type: object
      required:
      - id
      - email
      - username
      - role
      - status
      title: UserRead
    UserRole:
      type: string
      enum:
      - consumer
      - restaurant_manager
      - shelter_coordinator
      - driver
      - admin
      title: UserRole
    UserUpdate:
      properties:
        username:
          anyOf:
          - type: string
            maxLength: 80
            minLength: 3
          - type: 'null'
          title: Username
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      example:
        code: forbidden
        message: insufficient role
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT